follow unfollow

+++ app/Http/Controllers/FrontendController.php
@@ -259,6 +259,10 @@ class FrontendController extends Controller
     {
         $user = auth()->user();
         $isAlreadyFollowing = Follower::where('follower_id', $user->id)->where('followed_id', $request->followed_id)->exists();
+        if(auth()->user()->id === $request->followed_id){
+            // Prevent users from following themselves
+            return;
+        }
         if ($isAlreadyFollowing) {
             Follower::where('follower_id', $user->id)->where('followed_id', $request->followed_id)->delete();
         } else {
